+++ R/ard_categorical_ci.survey.design.R
@@ -34,7 +34,7 @@ ard_categorical_ci.survey.design <- function(data,
 
   # check inputs ---------------------------------------------------------------
   check_not_missing(data)
-  check_class(data, "survey.design")
+  check_class(data, c("survey.design", "svyrep.design"))
   check_not_missing(variables)
 
   cards::process_selectors(
@@ -67,6 +67,22 @@ ard_categorical_ci.survey.design <- function(data,
     value = value,
     ...
   )
+}
+
+# `svyrep.design` is a sibling class of `survey.design`, not a subclass, so it
+# needs its own method. The confidence intervals are computed by
+# `survey::svyciprop()`, which is not generic and handles replicate designs.
+# See #355.
+
+#' @rdname ard_categorical_ci.survey.design
+#' @export
+#' @examplesIf do.call(asNamespace("cardx")$is_pkg_installed, list(pkg = "survey"))
+#' # replicate-weight designs are also supported
+#' rclus1 <- survey::as.svrepdesign(dclus1)
+#'
+#' ard_categorical_ci(rclus1, variables = sch.wide)
+ard_categorical_ci.svyrep.design <- function(data, ...) {
+  ard_categorical_ci.survey.design(data = data, ...)
 }
 
 .calculate_ard_onesample_survey_ci <- function(FUN, data, variables, by, conf.level, value, ...) {