+++ R/ard_survey_svyttest.R
@@ -33,7 +33,7 @@ ard_survey_svyttest <- function(data, by, variables, conf.level = 0.95, ...) {
   check_not_missing(variables)
   check_not_missing(by)
   check_range(conf.level, range = c(0, 1))
-  check_class(data, cls = "survey.design")
+  check_class(data, cls = c("survey.design", "svyrep.design"))
   cards::process_selectors(data[["variables"]], by = {{ by }}, variables = {{ variables }})
   check_scalar(by)